879: refactor(openapi/rust): use paperclip as code generator r=tiagolobocastro a=tiagolobocastro

We can now stop using our fork and use paperclip which allows us to keep the templates locally on openapi/templates.

+++ control-plane/plugin/src/resources/volume.rs
@@ -200,7 +200,7 @@ impl Scale for Volume {
         output: &utils::OutputFormat,
     ) -> PluginResult {
         let req = openapi::models::ResizeVolumeBody {
-            size: requested_size as usize,
+            size: requested_size,
         };
         match RestClient::client()
             .volumes_api()
@@ -431,6 +431,6 @@ fn child_uuid(uri: &str) -> String {
     }
 }
 
-fn to_human_readable(val: isize) -> String {
-    ::utils::bytes::into_human(val as u64)
+fn to_human_readable(val: u64) -> String {
+    ::utils::bytes::into_human(val)
 }
